Move sandbox policy transforms into sandboxing crate
