+++ codex-rs/core/src/state/turn.rs
@@ -1,5 +1,6 @@
 //! Turn-scoped state and active turn metadata scaffolding.
 
+use codex_sandboxing::policy_transforms::merge_permission_profiles;
 use indexmap::IndexMap;
 use std::collections::HashMap;
 use std::sync::Arc;
@@ -19,7 +20,6 @@ use tokio::sync::oneshot;
 use crate::codex::TurnContext;
 use crate::protocol::ReviewDecision;
 use crate::protocol::TokenUsage;
-use crate::sandboxing::merge_permission_profiles;
 use crate::tasks::SessionTask;
 use codex_protocol::models::PermissionProfile;
