+++ codex-rs/sandboxing/Cargo.toml
@@ -16,6 +16,7 @@ codex-network-proxy = { workspace = true }
 codex-protocol = { workspace = true }
 codex-utils-absolute-path = { workspace = true }
 dirs = { workspace = true }
+dunce = { workspace = true }
 libc = { workspace = true }
 serde_json = { workspace = true }
 tracing = { workspace = true, features = ["log"] }
