+++ codex-rs/sandboxing/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod landlock;
 pub mod macos_permissions;
+pub mod policy_transforms;
 #[cfg(target_os = "macos")]
 pub mod seatbelt;
 #[cfg(target_os = "macos")]
